Refactor source folder
