+++ consensus/pow/sha256/sha256worker.go
@@ -1,4 +1,4 @@
-package consensus
+package sha256
 
 import (
     "bytes"
